api/requirements.txt
fastapi
pydantic>=2.0.0
uvicorn[standard]
psycopg2-binary
python-dotenv
requests
